Move Text and Pattern methods to Literal.cs
Refactored Rejigs by moving the Text and Pattern methods from Rejigs.cs to a new Literal.cs file for better separation of concerns. Merged Pattern tests into LiteralTests and deleted the redundant Pattern.cs test file.
